requirements.txt
fastapi 
redis-om
uvicorn[standard]
# "fastapi[standard]"
requests 
pymongo
pydantic-settings 
python-dotenv
motor
prometheus_client
pytest
pytest-mock



